Move visualization/dot.rs to top-level dot module
