+++ rust/rubydex/src/main.rs
@@ -10,7 +10,7 @@ use rubydex::{
         memory::MemoryStats,
         timer::{Timer, time_it},
     },
-    visualization::dot,
+    dot,
 };
 
 #[derive(Parser, Debug)]
