--- rust/rubydex/src/visualization.rs
@@ -1,6 +1,0 @@
-//! Graph visualization module for the Index structure.
-//!
-//! This module provides functionality to generate visual representations of the Index's
-//! internal graph structure, showing the relationships between Names, Definitions, and URIs.
-
-pub mod dot;
